fixed naming

+++ src/controllers/users/DeleteUserController.ts
@@ -1,16 +1,22 @@
-import { RequestHandler } from "express";
+import { Request, Response } from "express";
 import UserModel from "../../models/Users/UserModel";
 import ResponseService from "../../utils/ResponseService";
 import { TEXT } from "../../utils/JoiErrors";
+import { UserRequest } from "../../../test/utils/UserRequest";
 
-const deleteUserForAdminController: RequestHandler = async (req, res) => {
+const DeleteUserController = async (req: Request, res: Response) => {
   const userId = req.params.id;
   try {
     const prevUser = await UserModel.findOne({ _id: userId });
-
     if (!prevUser) {
       return ResponseService.error(res, TEXT.ERRORS.userDoesntExists);
     }
+
+    const jwtUserId = (req as UserRequest).userId;
+    if (jwtUserId !== req.params.id) {
+      return ResponseService.error(res, TEXT.ERRORS.somethingWentWrong);
+    }
+
     await UserModel.findOneAndDelete({ _id: userId });
 
     ResponseService.success(res, prevUser);
@@ -19,4 +25,4 @@ const deleteUserForAdminController: RequestHandler = async (req, res) => {
   }
 };
 
-export default deleteUserForAdminController;
+export default DeleteUserController;
